CBD-2753: Initial version of update_manifest_released program
A move of the 'add_to_released' program in build-tools/release to its own
repository, with some minor updates to naming.  Some initial preparation
for using Pipenv, but venv still in use in the build scripts until Pipenv
is installed on the relevant build systems and the full process to use
is in place.

Change-Id: I58b3def5b37c0a9dc19d67f642bdafeaadcc80dd
Reviewed-on: http://review.couchbase.org/101872
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>